requirements.txt
Markdown==2.4
PAM==0.4.2
Pillow==2.3.0
PyYAML==3.10
Pygments==1.6
RBTools==0.4.1
SecretStorage==2.0.0
Twisted-Core==13.2.0
Twisted-Web==13.2.0
Twisted-Words==13.2.0
adium-theme-ubuntu==0.3.4
apt-xapian-index==0.45
argparse==1.2.1
astroid==1.0.1
bzr==2.7.0dev1
chardet==2.0.1
colorama==0.2.5
command-not-found==0.3
configobj==4.7.2
debtagshw==0.1
decorator==3.4.0
defer==1.0.6
dirspec==13.10
dnspython==1.11.1
duplicity==0.6.23
git-review==1.23
html5lib==0.999
httplib2==0.8
keyring==3.5
launchpadlib==1.10.2
lazr.restfulclient==0.13.3
lazr.uri==1.0.3
lockfile==0.8
logilab-common==0.61.0
lxml==3.3.3
mercurial==2.8.2
numpy==1.8.1
oauth==1.0.1
oauthlib==0.6.1
oneconf==0.3.7
paramiko==1.10.1
pep8==1.4.6
pexpect==3.1
piston-mini-client==0.7.5
pyOpenSSL==0.13
pycrypto==2.6.1
pycups==1.9.66
pycurl==7.19.3
pyflakes==0.8.1
pygame==1.9.1release
pygobject==3.12.0
pygpgme==0.3
pyinotify==0.9.4
pylint==1.1.0
pyserial==2.6
pysmbc==1.0.14.1
python-apt==0.9.3.5
python-debian==0.1.21-nmu2ubuntu2
pyxdg==0.25
reportlab==3.0
requests==2.2.1
scipy==0.13.3
sessioninstaller==0.0.0
simplejson==3.3.1
six==1.5.2
software-center-aptd-plugins==0.0.0
ssh-import-id==3.21
system-service==0.1.6
unity-lens-photos==1.0
urllib3==1.7.1
urwid==1.1.1
virtualenv==1.11.4
wadllib==1.3.2
wsgiref==0.1.2
xdiagnose==3.6.3build2
zope.interface==4.0.5
